Moving `list_connected_cameras` into cameras file.
